ravu: boost the model with smoother kernel
First attempt to make RAVU softer.

On metrics, PSNR decreases by like `0.3dB` for `radius=3`. Training
set and test set is the same as the ones used for original model.
On the other hand, the visual difference is kind of subjective.

The second kernel training currently inherits the key from previous
kernel, which is probably a rather bad idea. But I couldn't figure out
a smarter approach at this point.